pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "muedit"
version = "2.0.0"
description = "FastICA-based Motor Unit decomposition of High-Density EMG signals"
readme = "README.md"
license = "GPL-3.0"
requires-python = ">=3.11"
authors = [
  { name = "Simon Avrillon" },
]
keywords = ["EMG", "motor unit", "decomposition", "ICA", "electromyography", "neuroscience"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
  "numpy>=1.26",
  "scipy>=1.13",
  "matplotlib>=3.9",
  "pandas>=2.2",
  "xmltodict>=0.14",
  "pydantic>=2.10",
  "PyYAML>=6.0",
  "h5py>=3.12",
  "fastapi>=0.115",
  "uvicorn>=0.32",
  "python-multipart>=0.0.12",
  "pyedflib>=0.1.36",
  "torch>=2.5",
]

[project.optional-dependencies]
dev = [
  "build>=1.2",
  "twine>=5.1",
  "pytest>=8.3",
  "ruff>=0.8",
]

[project.scripts]
muedit-api = "muedit.cli:serve_api"
muedit-decompose = "muedit.cli:run_decomposition_cli"

[tool.setuptools]
package-dir = {"" = "python/src"}

[tool.setuptools.packages.find]
where = ["python/src"]
include = ["muedit*", "adapt_decomp*"]

[project.urls]
Repository = "https://github.com/simonavrillon/MUedit2"
"Bug Tracker" = "https://github.com/simonavrillon/MUedit2/issues"
